move files + constants refactor start
